refactor(#N/A): extract methods and commands [guangchang]
