+++ src/types/methods.ts
@@ -1,0 +1,12 @@
+type ViewFileInitData = {
+  files: File[]
+  pinnedCids: string[]
+}
+
+type NodeInfos = {
+  'Node Status': string
+  'Peer ID': string
+  API: string
+  GateWay: string
+  'Public Key': string
+}
